Avoid using double in gamelogic code

+++ EpicRay/src/de/squareys/EpicRay/EpicRay/DynamicEntity.java
@@ -7,11 +7,10 @@ import de.squareys.EpicRay.Rendering.ISprite;
 
 public class DynamicEntity extends StaticEntity {
 
-	protected double m_vx;
-	protected double m_vy;
+	protected float m_vx;
+	protected float m_vy;
 	
-	protected double m_speed;
-	
+	protected float m_speed;
 	
 	public DynamicEntity(int p_x, int p_y, int p_w, int p_h, ISprite sprite, IWorld world) {
 		super(p_x, p_y, p_w, p_h, sprite, world);
@@ -26,8 +25,8 @@ public class DynamicEntity extends StaticEntity {
 
 	@Override
 	public void onUpdate() {
-		double oldX = m_x;
-		double oldY = m_y;
+		float oldX = m_x;
+		float oldY = m_y;
 		
 		m_x += m_vx * m_speed;
 		m_y += m_vy * m_speed;
@@ -44,11 +43,11 @@ public class DynamicEntity extends StaticEntity {
 		}
 	}
 	
-	public void setSpeed(double speed){
+	public void setSpeed(float speed){
 		m_speed = speed;
 	}
 	
-	public void setMovementDir(double vx, double vy){
+	public void setMovementDir(float vx, float vy){
 		m_vx = vx;
 		m_vy = vy;
 	}
